test(i18n): cover fillLocales
