+++ packages/i18n/src/locales.ts
@@ -16,3 +16,19 @@ export const locales = LOCALES;
 export function resolveLocale(value: string | undefined): Locale {
   return locales.includes(value as Locale) ? (value as Locale) : "en";
 }
+
+/**
+ * Ensure all locales have a value, filling in missing entries with a fallback.
+ */
+export function fillLocales(
+  values: Partial<Record<Locale, string>> | undefined,
+  fallback: string
+): Record<Locale, string> {
+  return LOCALES.reduce<Record<Locale, string>>(
+    (acc, locale: Locale) => {
+      acc[locale] = values?.[locale] ?? fallback;
+      return acc;
+    },
+    {} as Record<Locale, string>
+  );
+}
