+++ packages/platform-core/__tests__/createShopUtils.test.ts
@@ -2,7 +2,7 @@
 import fs from "fs";
 import { copyTemplate } from "../src/createShop/fsUtils";
 import { loadBaseTokens } from "../src/createShop/themeUtils";
-import { fillLocales } from "../src/utils/locales";
+import { fillLocales } from "@i18n/locales";
 
 describe("createShop utils", () => {
   beforeEach(() => {
