+++ packages/platform-core/src/utils/README.md
@@ -4,8 +4,6 @@ Shared helpers for platform-core.
 
 - `getShopFromPath(path)` – extract a shop identifier from a URL path.
 - `replaceShopInPath(path, shop)` – swap the shop identifier in a path.
-- `fillLocales(values, fallback)` – ensure all supported locales have a value,
-  falling back when missing.
 
 These utilities are re-exported from the `@acme/platform-core` package root for
 easy consumption. Functions like `slugify` and `genSecret` now live in
