+++ packages/platform-core/src/utils/index.ts
@@ -1,3 +1,2 @@
 export { getShopFromPath } from "./getShopFromPath";
 export { replaceShopInPath } from "./replaceShopInPath";
-export { fillLocales } from "./locales";
